Update OTLP receiver configuration and endpoint in manifests
The OTLP receiver configuration in the app-otel.yaml manifest has been updated, with specifications for gRPC and HTTP protocols and the addition of hostmetric scrapers. The HTTP endpoint reference for the OTLP in the app-tempo.yaml manifest has also been updated. These changes contribute to improving the collection and transmission of observability data.